+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add information about Node Pools and Worker Nodes in a cluster.
+
 ## [2.8.0] - 2021-09-06
 
 ### Added